+++ src/solver/tableau/dp54.rs
@@ -1,5 +1,8 @@
 //! Dormand–Prince Method
-pub const RK_ORDER: usize = 5;
+
+#![allow(dead_code)]
+
+pub const RK_ORDER: i32 = 5;
 pub const RK_DIM: usize = 7;
 pub const RK_A: [[f64; RK_DIM - 1]; RK_DIM - 1] = [
     [1.0 / 5.0, 0.0, 0.0, 0.0, 0.0, 0.0],
